+++ changelog.md
@@ -4,6 +4,7 @@
 ### features
 - add open window label on dock for app items (configurable).
 - add input to write custom text items on the toolbar.
+- add export resource button on developer tools.
 
 ### fix
 - not removing old icons mask on icon pack change.